Move AssayUploadTabs to assay/constants.ts
Circular imports were causing test failures.
