+++ packages/components/src/internal/components/assay/RunDataPanel.spec.tsx
@@ -8,7 +8,7 @@ import { gridInit } from '../../actions';
 import { FormTabs, withFormSteps, WithFormStepsProps } from '../../..';
 import { ASSAY_WIZARD_MODEL } from '../../../test/data/constants';
 import { initUnitTestMocks } from '../../testHelperMocks';
-import { AssayUploadTabs } from '../../AssayDefinitionModel';
+import { AssayUploadTabs } from '../../constants';
 
 import { RunDataPanel } from './RunDataPanel';
